ceviche_engine/requirements.txt
feedparser>=6.0.10
requests>=2.31.0
beautifulsoup4>=4.12.0
python-frontmatter>=1.0.0
python-dateutil>=2.8.2
